pyproject.toml
[project]
name = "ApplicatorAI"
version = "0.1.3"
description = "An AI-powered cover letter generator with multiple features"
authors = [
    {name = "Subhash B S", email = "subhashbs36@gmail.com"},
]
dependencies = [
    "gradio==4.19.2",
    "fpdf2==2.7.6",
    "PyPDF2==3.0.1",
    "docx2txt==0.8",
    "reportlab==4.0.9",
    "gradio-pdf==0.0.1",
    "beautifulsoup4==4.12.3",
    "requests==2.31.0",
    "google-generativeai==0.3.2",
    "python-dotenv==1.0.1",
    "markdown==3.5.2",
    "aiofiles>=22.0,<24.0",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true 